requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.5
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
